pyproject.toml
[project]
name = "oz-django8"
version = "0.1.0"
description = ""
authors = [
    {name = "hak1319",email = "gkr054@naver.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "cryptography (>=44.0.1,<45.0.0)",
    "django-ninja (>=1.3.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.0"
mypy = "^1.15.0"
django-stubs = "^5.1.3"
types-pymysql = "^1.1.0.20241103"
coverage = "^7.6.12"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
python_version = "3.13"
strict = true

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "manage"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "oz_django8.settings"